+++ R/geom_mlb_headshots.R
@@ -13,8 +13,7 @@
 #'   \item{`x`}{ - The x-coordinate. Required.}
 #'   \item{`y`}{ - The y-coordinate. Required.}
 #'   \item{`player_id`}{ - The players' MLBAM (Baseball Savant) id. Required.}
-#'   \item{`na_headshot_to_logo = TRUE`}{ - Should NA/non-matches return the MLB logo
-#'                                          instead of a grayed out blank headshot? Only used with `geom_mlb_headshots()`. Defaults to `TRUE`}
+
 #'   \item{`alpha = NULL`}{ - The alpha channel, i.e. transparency level, as a numerical value between 0 and 1.}
 #'   \item{`colour = NULL`}{ - The image will be colorized with this colour. Use the special character `"b/w"` to set it to black and white. For more information on valid colour names in ggplot2 see <https://ggplot2.tidyverse.org/articles/ggplot2-specs.html?q=colour#colour-and-fill>}
 #'   \item{`angle = 0`}{ - The angle of the image as a numerical value between 0° and 360°.}
@@ -45,28 +44,28 @@
 #' df <- data.frame(
 #'   a = c(rep(1:3, 3), 1.5, 2.5),
 #'   b = c(sort(rep(1:3, 3), decreasing = TRUE), 2.5, 2.5),
-#'   player_id = c("660670",
-#'                 "545361",
-#'                 "605141",
-#'                 "571448",
-#'                 "594798",
+#'   player_id = c("665742",
+#'                 "677594",
+#'                 "592450",
+#'                 "608070",
+#'                 "694973",
 #'                 "518692",
 #'                 "0",
-#'                 "521692",
+#'                 "663728",
 #'                 "120074",
-#'                 "665487",
-#'                 "518934"),
-#'   player_name = c("Ronald Acuña Jr.",
-#'                   "Mike Trout",
-#'                   "Mookie Betts",
-#'                   "Nolan Arenado",
-#'                   "Jacob deGrom",
+#'                 "608369",
+#'                 "606466"),
+#'   player_name = c("Juan Soto",
+#'                   "Julio Rodríguez",
+#'                   "Aaron Judge",
+#'                   "José Ramírez",
+#'                   "Paul Skenes",
 #'                   "Freddie Freeman",
 #'                   "Non Match",
-#'                   "Salvador Perez",
+#'                   "Cal Raleigh",
 #'                   "David Ortiz",
-#'                   "Fernando Tatis Jr.",
-#'                   "DJ LeMahieu")
+#'                   "Corey Seager",
+#'                   "Ketel Marte")
 #' )
 #'
 #' # set a custom fill colour for one player
@@ -79,15 +78,14 @@
 #'   coord_cartesian(xlim = c(0.75, 3.25), ylim = c(0.7, 3.25)) +
 #'   theme_void()
 #'
-#' # apply alpha as constant and use non default na replacement
+#' # apply alpha as constant
 #' ggplot(df, aes(x = a, y = b)) +
-#'   geom_mlb_headshots(aes(player_id = player_id), height = 0.2, alpha = 0.5,
-#'                      na_headshot_to_logo = FALSE) +
+#'   geom_mlb_headshots(aes(player_id = player_id), height = 0.2, alpha = 0.5) +
 #'   geom_label(aes(label = player_name), nudge_y = -0.35, alpha = 0.5) +
 #'   coord_cartesian(xlim = c(0.75, 3.25), ylim = c(0.7, 3.25)) +
 #'   theme_void()
 #'
-#' # apply colour as an aesthetic and use the dot version
+#' # apply colour as an aesthetic and use the dot version with colored background
 #' ggplot(df, aes(x = a, y = b)) +
 #'   geom_mlb_dot_headshots(aes(player_id = player_id, colour = colour), height = 0.2) +
 #'   geom_label(aes(label = player_name), nudge_y = -0.35, alpha = 0.5) +
@@ -143,18 +141,12 @@ GeomMLBheads <- ggplot2::ggproto(
   # non_missing_aes = c(""),
   default_aes = ggplot2::aes(
     alpha = NULL, colour = NULL, angle = 0, hjust = 0.5,
-    vjust = 0.5, width = 1.0, height = 1.0, na_headshot_to_logo = TRUE
+    vjust = 0.5, width = 1.0, height = 1.0
   ),
   draw_panel = function(data, panel_params, coord, na.rm = FALSE, na_headshot_to_logo = TRUE) {
     data <- coord$transform(data, panel_params)
 
-    if (na_headshot_to_logo) {
-      grobs <- lapply(seq_along(data$player_id), build_grobs, alpha = data$alpha, colour = data$colour, data = data, type = "logo_headshots")
-    } else {
-      grobs <- lapply(seq_along(data$player_id), build_grobs, alpha = data$alpha, colour = data$colour, data = data, type = "gray_headshots")
-    }
-
-
+    grobs <- lapply(seq_along(data$player_id), build_grobs, alpha = data$alpha, colour = data$colour, data = data, type = "transparent_dot_headshots")
 
     class(grobs) <- "gList"