+++ R/mlb_tier_maker.R
@@ -234,8 +234,8 @@ mlb_team_tiers <- function(data,
 #' \donttest{
 #' library(ggplot2)
 #' library(dplyr, warn.conflicts = FALSE)
-#' player_ids <- load_headshots() |>
-#'   head(35) |>
+#' player_ids <- load_headshots() %>%
+#'   head(35) %>%
 #'   pull(savant_id)
 #'
 #' # Build the player tiers data frame